上传增加Cluster功能

+++ src/proto/build.sh
@@ -1,8 +1,14 @@
 #!/bin/sh
 
-rm gameproto/*
+
+gen_files=`ls gameproto/`
+
+if [ ! -z "$gen_files" ]
+then
+	rm gameproto/*
+fi
 
 cd conf/
 
-../protoc --plugin=../protoc-gen-go --go_out=../gameproto/ *proto
+protoc --plugin=../protoc-gen-go --go_out=../gameproto/ *proto
